requirements.txt
ipykernel==6.30.1
pandas==2.3.2
numpy==1.26.4

opencv-python==4.10.0.84
matplotlib==3.10.6

scikit-learn==1.7.2
torch==2.3.1
torchvision==0.18.1

seaborn>=0.13.0
plotly>=5.0.0

tqdm>=4.66.0
tensorboard>=2.8.0

pillow>=10.0.0
streamlit>=1.32.0

